refactor: CompanyClient 패키지 변경에 따른 import 경로 수정
