+++ src/main/java/com/yeoljeong/tripmate/product/application/client/CompanyClient.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.product.application.service.client;
+package com.yeoljeong.tripmate.product.application.client;
 
 import com.yeoljeong.tripmate.product.infrastructure.external.dto.CompanyClientResponse;
 import java.util.UUID;
